fix: exporting in OTLP example
Shutting down the tracer provider via the opentelemetry::global API
has not worked since opentelemetry 0.24 when the batch processor
is in use: buffered traces were not exported. It's not necessary
to install the global provider for the tracing integration, as tracing
has its own way to manage the global subscriber.
Instead, manage the instance in the scope guard the same way as the
meter provider instance, and call the shutdown method on it when the
guard is dropped.